Going to redefine queue

+++ queue.h
@@ -1,6 +1,7 @@
 #ifndef QUEUE_H
 #define QUEUE_H
 #include <stdbool.h>
+#include <stdint.h>
 
 typedef struct List {
 	unsigned long long val;
@@ -14,11 +15,11 @@ typedef struct Queue {
 	int num_elements;
 } Queue;
 
-extern void push(Queue* q, unsigned long long val);
+extern void push(Queue* q, uint64_t val);
 
-extern unsigned long long top(Queue* q);
+extern uint64_t top(Queue* q);
 
-extern unsigned long long pop(Queue* q);
+extern uint64_t pop(Queue* q);
 
 extern void deleteQueue(Queue* q);
